Moved a bunch of stuff inside folders
call 1/10 done
