+++ src/routes/app/call/+page.svelte
@@ -1,0 +1,6 @@
+<script>
+	import { goto } from '$app/navigation';
+	import { onMount } from 'svelte';
+
+	onMount(() => goto(`/app/dashboard`));
+</script>
