+++ src/routes/app/call/[targetId]/+page.svelte
@@ -1,0 +1,9 @@
+<script>
+	import Call from '$lib/communication/Call.svelte';
+	import { page } from '$app/stores';
+
+
+	$: targetId = $page.params.targetId; 
+</script>
+
+<Call {targetId} />
