--- src/routes/app/contacts/call/+page.svelte
@@ -1,5 +1,0 @@
-<script>
-	import Call from '$lib/call/Call.svelte';
-</script>
-
-<Call />
