+++ src/routes/app/music/+page.svelte
@@ -130,7 +130,6 @@
 					/>
 				</div>
 			</div>
-			
 		</div>
 	</div>
 </main>
